main: suggested repositories
I suggest to use PSR repositories whenever possible. The PHP Standard Recommendation (PSR) is a PHP specification published by the PHP Framework Interop Group. PSR serves the standardization of programming concepts in PHP. The aim is to enable interoperability of components and to provide a common technical basis for implementation of proven concepts for optimal programming and testing practices.